lint: prepare prod build

+++ pages/auth/logout.tsx
@@ -5,9 +5,10 @@ export default function Logout() {
   const router = useRouter();
 
   React.useEffect(() => {
+    if (!router.isReady) return;
     localStorage.clear();
     router.push('/auth/login');
-  }, []);
+  }, [router]);
 
   return 'logging out...';
 }
